requirements.txt
streamlit==1.52.0
qdrant-client==1.11.0
minio==7.2.8
google-generativeai==0.7.0
sentence-transformers==3.1.1
pyvi==0.1.1
python-docx==1.1.2
tqdm==4.66.5
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
Pillow>=10.0.0